fix(scanner): cover legacy backup fencing gaps
