+++ crates/scanner/src/scanner/leadership.rs
@@ -273,8 +273,9 @@ pub(super) async fn fence_scanner_usage_epoch_with_expected_epoch(
             Some(epoch) if epoch == claimed_epoch => return Ok(()),
             Some(_) | None => {}
         }
-        // A validated pre-marker baseline needs its explicit identity before
-        // acquiring an epoch, which makes it ineligible for legacy decoding.
+        // A validated pre-marker legacy baseline needs an explicit complete
+        // identity before acquiring an epoch. Otherwise the v2 reader would
+        // reject the fenced value on its next startup.
         if !usage.usage_snapshot_bootstrap_pending {
             usage.usage_snapshot_complete = true;
         }
